+++ www/wp-content/themes/hittheroad-2021/woocommerce/myaccount/form-reset-password.php
@@ -12,7 +12,7 @@
  *
  * @see https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 7.0.1
+ * @version 9.2.0
  */
 
 defined('ABSPATH') || exit; ?>
@@ -25,13 +25,13 @@ defined('ABSPATH') || exit; ?>
 		<p><?= apply_filters('woocommerce_reset_password_message', esc_html__('Enter a new password below.', 'woocommerce')); ?></p><?php // @codingStandardsIgnoreLine ?>
 
 		<div class="form-floating mb-3">
-			<label for="password_1"><?php esc_html_e('New password', 'woocommerce'); ?>&nbsp;<span class="required">*</span></label>
-			<input type="password" class="woocommerce-Input woocommerce-Input--text input-text" name="password_1" id="password_1" autocomplete="new-password">
+			<label for="password_1"><?php esc_html_e('New password', 'woocommerce'); ?> <span class="required" aria-hidden="true">*</span><span class="screen-reader-text"><?php esc_html_e('Required', 'woocommerce'); ?></span></label>
+			<input type="password" class="woocommerce-Input woocommerce-Input--text input-text" name="password_1" id="password_1" autocomplete="new-password" required aria-required="true">
 		</div>
 
 		<div class="form-floating mb-3">
-			<label for="password_2"><?php esc_html_e('Re-enter new password', 'woocommerce'); ?>&nbsp;<span class="required">*</span></label>
-			<input type="password" class="woocommerce-Input woocommerce-Input--text input-text" name="password_2" id="password_2" autocomplete="new-password">
+			<label for="password_2"><?php esc_html_e('Re-enter new password', 'woocommerce'); ?> <span class="required" aria-hidden="true">*</span><span class="screen-reader-text"><?php esc_html_e('Required', 'woocommerce'); ?></span></label>
+			<input type="password" class="woocommerce-Input woocommerce-Input--text input-text" name="password_2" id="password_2" autocomplete="new-password" required aria-required="true">
 		</div>
 
 		<input type="hidden" name="reset_key" value="<?= esc_attr($args['key']); ?>">